requirements.txt
validators==0.20.0
python-dotenv==0.21.0
Google-Images-Search==1.4.6
wheel==0.38.4
crispy-bootstrap5==0.6
Django==4.1.1
django-allauth==0.51.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.4
whitenoise==6.2.0
psycopg2-binary==2.9.3
